fix(usenet): parse cooldown config as long and mark the docs version
Validation for these two keys accepts the whole Int64 range, so an int
parse treated an accepted out-of-range value as unset and returned the
default rather than clamping it. A value like 2147483648 passed both the
Settings API and startup validation, then resolved to 60s instead of the
documented 300s ceiling. Both getters now parse long and clamp before
converting, with boundary tests either side of Int32.

Also adds the release marker to the new docs heading, since older images
reject the new config names at startup and readers need the minimum
version.

+++ backend/Config/ConfigManager.cs
@@ -1051,9 +1051,9 @@ public class ConfigManager : IConfigReader, IConfigUpdater, IConfigChangeSource
     public TimeSpan GetCircuitBreakerInitialCooldown()
     {
         var configured = StringUtil.EmptyToNull(GetConfigValue(ConfigKeys.UsenetCircuitBreakerInitialCooldownSeconds));
-        if (configured is null || !int.TryParse(configured, out var seconds))
+        if (configured is null || !long.TryParse(configured, out var seconds))
             return TimeSpan.FromSeconds(60);
-        return TimeSpan.FromSeconds(Math.Clamp(seconds, 5, 300));
+        return TimeSpan.FromSeconds(Math.Clamp(seconds, 5L, 300L));
     }
 
     /// <summary>
@@ -1065,9 +1065,9 @@ public class ConfigManager : IConfigReader, IConfigUpdater, IConfigChangeSource
     public TimeSpan GetCircuitBreakerMaxCooldown()
     {
         var configured = StringUtil.EmptyToNull(GetConfigValue(ConfigKeys.UsenetCircuitBreakerMaxCooldownSeconds));
-        if (configured is null || !int.TryParse(configured, out var seconds))
+        if (configured is null || !long.TryParse(configured, out var seconds))
             return TimeSpan.FromSeconds(300);
-        return TimeSpan.FromSeconds(Math.Clamp(seconds, 5, 3600));
+        return TimeSpan.FromSeconds(Math.Clamp(seconds, 5L, 3600L));
     }
 
     /// <summary>
